Add stylus button support

+++ engine/src/flutter/shell/platform/windows/flutter_window.h
@@ -93,7 +93,7 @@ class FlutterWindow : public KeyboardManager::WindowDelegate,
                              double y,
                              FlutterPointerDeviceKind device_kind,
                              int32_t device_id,
-                             UINT button,
+                             uint64_t button,
                              uint32_t rotation,
                              uint32_t pressure);
 
@@ -103,7 +103,7 @@ class FlutterWindow : public KeyboardManager::WindowDelegate,
                            double y,
                            FlutterPointerDeviceKind device_kind,
                            int32_t device_id,
-                           UINT button);
+                           uint64_t button);
 
   // Called when the mouse leaves the window.
   virtual void OnPointerLeave(double x,
